currentcolor support for color-mix
color-mix values are generally resolved into single blink::Colors at
computed value time. "currentcolor" cannot do this as its value is
unknown until after style is computed.

In the event that a color-mix contains "currentcolor" this CL adds
a private UnresolvedColorMix struct to StyleColor that will get
resolved at used value time when currentcolor is available. The
internal color in a StyleColor can now either be a blink::Color
or a pointer to an UnresolvedColorMix.

If neither color is a "currentcolor" (or a color-mix value with a
"currentcolor") we continue to resolve color-mix values at computed value time.

It is still unresolved what we'll do when the value of both percentages
for color-mix are zero as it should get rejected at parse time but
there are no WPT tests that cover this situation yet.

Changes to third_party/blink/web_tests/external/wpt/css/css-color/parsing/color-computed-color-mix-function-expected.txt are due
to new tests passing. New failures are all old failures just the
message has changed.

Bug: 1333988
Change-Id: I954f2371bc267e7912b62dbaccc363e5bdf901da