[ComputePressure] Add Permissions integration
Make ComputePressure controlled by Permissions according to the
spec [1]. However, there should be no permission prompt and it should
auto grant. Because there is not a reasonable way to ask the user and
expect them to understand. It also doesn't need to have an entry for
Compute Pressure in chrome://settings/content.

[1] https://wicg.github.io/compute-pressure/#permissions-and-user-prompts

Bug: 1374200
Change-Id: Ib5f8e8a067ceafcfaf2c157d47e81ea858331ddc