Fix NavigationParams::origin_agent_cluster for about:blank.
When navigating to about:blank, the page "inherits" the origin of the
navigation initiator. Thus, an about:blank page may very well be assigned to an origin-agent cluster, despite it not being an HTTP page with HTTP headers to request origin-agent clustering.

This CL ensures NavigationRequest::DetermineOriginAgentClusterEndResult is run for about:blank.

Bug: 1399759
Change-Id: Iec6e0a3198c5f2f46498dca9541ed250bfc1b10d