[scroll-animations] The animation shorthand should set timeline to auto
We currently allow the timeline to be specified as part of the
animation shorthand. This is (for now) incorrect per a recent
CSSWG resolution [1].

This CL changes all WPTs to use the animation-timeline longhand
instead. I've also added a setup assertion (or equivalent) to many
tests to prevent UAs without support for animation-timeline from
running the animations on the document timeline (which could lead
to unstable/timing-dependent failure results).

[1] https://github.com/w3c/csswg-drafts/issues/6946#issuecomment-1233190360

Fixed: 1397870
Change-Id: I97cd13481ea00a074f50c132562d9bb83e7bc4dd