Use ComputedStyle to determine CQ containers in ancestors
StyleRecalcContext:From[Inclusive]Ancestors currently look
at an ancestor's ContainerQueryEvaluator to determine whether
or not that ancestor is a CQ container. This is correct for
normal circumstances where style recalc has happened for the
ancestor, but not if the ancestor is in display:none.
In that case, the ancestor is still a container, but there
will be no evaluator.

This CL fixes this by instead checking the ComputedStyle
of the ancestor to determine containership.

Fixed: 1399755
Change-Id: Ia4096d158737c2e98a9215a3d8e7b2273d77320f