VT: Use geometry mapper for ancestor rect mapping.
The non-geometry mapper path seems to not work correctly for fragmented
containers (producing a value of 0x0, filed crbug.com/1396368). This
causes us to determine wrong overflow values. The fix is to use a
geometry mapper, which is best practice anyway.

This patch also rewrites the function to add extra conditions necessary
for the overflow calculation to be correct

R=bokan@chromium.org, wangxianzhu@chromium.org

Bug: 1393064, 1399538
Change-Id: I40063eb1d0418c86e0405b48714ba033ef58726e